% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categMiscRate.R
\name{categMisclassRates}
\alias{categMisclassRates}
\title{Categorical misclassification rates}
\usage{
categMisclassRates(predIntervalMatrix, iprint = FALSE)
}
\arguments{
\item{predIntervalMatrix}{A contingency table (matrix or data frame) where rows
represent true classes and columns represent predicted intervals.}

\item{iprint}{Logical; if \code{TRUE}, prints details for each class during computation.}
}
\value{
A matrix with misclassification rates and average prediction interval lengths for
each category, along with the overall average for each metric.
}
\description{
This function computes the misclassification rates and average lengths
of prediction intervals for each category based on the provided prediction interval matrix.
}
\examples{
set.seed(1)
# Example data
predIntervalMatrix <- matrix(sample(0:5, 12, replace = TRUE), nrow = 3, ncol = 4)
rownames(predIntervalMatrix) <- c("A", "B", "C")
colnames(predIntervalMatrix) <- c("A", "A,B", "B,C", "C")

# Compute misclassification rates and average interval lengths
result <- categMisclassRates(predIntervalMatrix, iprint = FALSE)
print(result)

}
