% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logscore.R
\name{logscore}
\alias{logscore}
\title{Negative log-likelihood score for classification}
\usage{
logscore(predmatrix, labels)
}
\arguments{
\item{predmatrix}{A numeric matrix where each row contains predicted probabilities for each class.}

\item{labels}{A numeric vector of true class indices corresponding to each row of predmatrix.}
}
\value{
The negative log-likelihood score (numeric).
}
\description{
Calculate the negative log-likelihood score for classification.
}
