% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partCor.R
\name{partCor}
\alias{partCor}
\title{Partial correlation estimation}
\usage{
partCor(S, given, j, k)
}
\arguments{
\item{S}{A correlation matrix.}

\item{given}{A vector of indices for the conditioning variables.}

\item{j}{An index for the first variable.}

\item{k}{An index for the second variable.}
}
\value{
A numeric value representing the partial correlation between \code{j} and \code{k}
given the variables in \code{given}.
}
\description{
Computes the partial correlation between two variables, \code{j} and \code{k},
given a set of conditioning variables using the correlation matrix.
}
\examples{
# Example correlation matrix
S <- matrix(c(1, 0.5, 0.3, 0.5, 1, 0.2, 0.3, 0.2, 1), ncol = 3)
# Partial correlation of variables 1 and 3 given variable 2
partCor(S, given = 2, j = 1, k = 3)

}
